Move API to EEB site
