+++ sites/rrmarketplaceblog.com/server/graphql/queries/equipment-exchange-content.js
